# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Deployer API",
    "description": "REST API for deploying containers from templates. Async: POST deploy/lifecycle → 202 + operation, poll GET /api/operations/:id. Deploy: containerName + params.",
    "version": "2.0.0"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "session": { "type": "apiKey", "in": "cookie", "name": "s" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" },
      "mcpBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "MCP key dep_mcp_live_… (full Deployer API via /mcp tools)"
      }
    },
...
  },
  "security": [{ "session": [] }, { "apiKey": [] }],
  "paths": {
...
    "/api/v1/mcp/keys": {
      "get": {
        "summary": "List MCP keys (session only)",
        "responses": { "200": { "description": "Keys list with mcpUrl" }, "401": { "description": "Unauthorized" } },
        "security": [{ "session": [] }]
      },
      "post": {
        "summary": "Issue MCP key (plaintext once, max 5)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "label": { "type": "string" } },
                "required": ["label"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Key created with plaintext" },
          "400": { "description": "mcp_key_limit" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "session": [] }]
      }
    },
    "/api/v1/mcp/keys/{keyId}/revoke": {
      "post": {
        "summary": "Revoke MCP key",
        "parameters": [{ "name": "keyId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Revoked" }, "404": { "description": "Not found" } },
        "security": [{ "session": [] }]
      }
    },
...
  }
}